+++ src/client/App.tsx
@@ -87,7 +87,7 @@ import ZaakStatus from './pages/ZaakStatus/ZaakStatus';
 import ZorgThemaPagina from './pages/Zorg/Zorg';
 import ZorgVoorzieningen from './pages/Zorg/ZorgRegelingen';
 import ZorgDetail from './pages/ZorgDetail/ZorgDetail';
-import { useMonitoring } from './utils/monitoring';
+import { useMonitoring } from './helpers/monitoring';
 
 function AppNotAuthenticated() {
   useSetDeeplinkEntry(['sso', 'authMethod']);
