+++ src/client/components/MyNotifications/Notification.tsx
@@ -17,8 +17,8 @@ import { ThemaTitles } from '../../config/thema';
 import { trackItemClick } from '../../hooks/analytics.hook';
 import { useContentDimensions } from '../../hooks/useContentDimensions';
 import { useProfileTypeValue } from '../../hooks/useProfileType';
-import { trackEvent } from '../../utils/monitoring';
-import { isInteralUrl } from '../../utils/utils';
+import { trackEvent } from '../../helpers/monitoring';
+import { isInteralUrl } from '../../helpers/utils';
 import Linkd from '../Button/Button';
 import { DocumentLink } from '../DocumentList/DocumentLink';
 import ThemaIcon from '../ThemaIcon/ThemaIcon';
