+++ src/client/hooks/api/useDataApi.ts
@@ -4,7 +4,7 @@ import axios, { AxiosRequestConfig, AxiosResponseTransformer } from 'axios';
 
 import { apiErrorResult } from '../../../universal/helpers/api';
 import { Action } from '../../../universal/types';
-import { captureException } from '../../utils/monitoring';
+import { captureException } from '../../helpers/monitoring';
 
 export interface ApiRequestOptions extends AxiosRequestConfig {
   postpone?: boolean;
