+++ src/client/hooks/useTrackThemas.hook.test.tsx
@@ -6,7 +6,7 @@ import { MutableSnapshot, RecoilRoot } from 'recoil';
 import { appStateAtom } from './useAppState';
 import { useTrackThemas } from './useTrackThemas.hook';
 import { AppState } from '../../universal/types';
-import { trackEvent } from '../utils/monitoring';
+import { trackEvent } from '../helpers/monitoring';
 
 vi.mock('../utils/monitoring', () => ({
   trackEvent: vi.fn(),
