+++ src/client/hooks/useUsabilla.ts
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { usePhoneScreen } from './media.hook';
 import { useScript } from './useScript';
 import { IS_AP } from '../../universal/config/env';
-import { captureException } from '../utils/monitoring';
+import { captureException } from '../helpers/monitoring';
 
 const MAX_WAIT_FOR_USABILA_LIVE_MS = 5000; // 5 seconds
 const USABILLA_ID_MOBILE = '9fd5da44aa5b';
