+++ src/client/pages/NotFound/NotFound.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import { PageContent, PageHeading, TextPage } from '../../components';
-import { captureMessage } from '../../utils/monitoring';
+import { captureMessage } from '../../helpers/monitoring';
 
 export default function NotFound() {
   const location = useLocation();
